cloud_track/azure/assignment2/requirements.txt
azure-identity
azure-mgmt-storage
azure-storage-blob
python-dotenv